ai-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
transformers==4.35.2
torch==2.1.1
torchvision==0.16.1
torchaudio==2.1.1
sentence-transformers==2.2.2
pytesseract==0.3.10
Pillow==10.1.0
PyPDF2==3.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
numpy==1.24.4
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
